+++ src/wordless_colligation.py
@@ -1147,7 +1147,11 @@ def generate_table(main, table):
             settings['search_settings']['multi_search_mode'] and settings['search_settings']['search_terms']):
             dialog_progress = wordless_dialog_misc.Wordless_Dialog_Progress_Process_Data(main)
 
-            worker_colligation_table = Wordless_Worker_Colligation_Table(main, dialog_progress, update_gui)
+            worker_colligation_table = Wordless_Worker_Colligation_Table(
+                main,
+                dialog_progress = dialog_progress,
+                update_gui = update_gui
+            )
             thread_colligation_table = wordless_threading.Wordless_Thread(worker_colligation_table)
 
             thread_colligation_table.start()
@@ -1271,7 +1275,11 @@ def generate_fig(main):
             settings['search_settings']['multi_search_mode'] and settings['search_settings']['search_terms']):
             dialog_progress = wordless_dialog_misc.Wordless_Dialog_Progress_Process_Data(main)
 
-            worker_colligation_fig = Wordless_Worker_Colligation_Fig(main, dialog_progress, update_gui)
+            worker_colligation_fig = Wordless_Worker_Colligation_Fig(
+                main,
+                dialog_progress = dialog_progress,
+                update_gui = update_gui
+            )
             thread_colligation_fig = wordless_threading.Wordless_Thread(worker_colligation_fig)
 
             thread_colligation_fig.start()