+++ src/CHANGELOG.md
@@ -43,7 +43,7 @@
 
 ### ⏫ Dependency Upgrades
 - Upgrade Beautiful Soup to 4.8.1
-- Upgrade botok (pybo) to 0.6.10
+- Upgrade botok (pybo) to 0.6.18
 - Upgrade cChardet to 2.1.5
 - Upgrade Matplotlib to 3.1.2
 - Upgrade NetworkX to 2.4